refactor(fxa-payments-server): Create PaymentConsentCheckbox component
- refs #9361
- was able to componentize the checkbox from the payment form
- consolidated tests to PaymentConsentCheckbox/index.test.jsx
