+++ packages/fxa-payments-server/src/routes/Product/SubscriptionUpgrade/index.test.tsx
@@ -5,7 +5,7 @@ import '@testing-library/jest-dom/extend-expect';
 
 import { APIError } from '../../../lib/apiClient';
 import { Plan } from '../../../store/types';
-import { PlanUpgradeDetails, PlanDetailsCard } from './PlanUpgradeDetails';
+import { PlanDetailsCard } from './PlanUpgradeDetails';
 
 jest.mock('../../../lib/sentry');
 
@@ -33,14 +33,6 @@ import { SignInLayout } from '../../../components/AppLayout';
 
 import SubscriptionUpgrade, { SubscriptionUpgradeProps } from './index';
 import { getLocalizedCurrency } from '../../../lib/formats';
-
-const findMockPlan = (planId: string): Plan => {
-  const plan = MOCK_PLANS.find((x) => x.plan_id === planId);
-  if (plan) {
-    return plan;
-  }
-  throw new Error('unable to find suitable Plan object for test execution.');
-};
 
 describe('routes/Product/SubscriptionUpgrade', () => {
   afterEach(() => {
@@ -145,239 +137,6 @@ describe('routes/Product/SubscriptionUpgrade', () => {
     fireEvent.click(getByTestId('confirm'));
     expect(updateSubscriptionPlanMounted).toBeCalledTimes(1);
     expect(updateSubscriptionPlanEngaged).toBeCalledTimes(1);
-  });
-
-  describe('Legal', () => {
-    describe('rendering the legal checkbox Localized component', () => {
-      const baseProps = {
-        isMobile: false,
-        profile: PROFILE,
-        customer: CUSTOMER,
-        upgradeFromPlan: UPGRADE_FROM_PLAN,
-        upgradeFromSubscription: CUSTOMER.subscriptions[0],
-        updateSubscriptionPlanStatus: {
-          error: null,
-          loading: false,
-          result: null,
-        },
-        updateSubscriptionPlanAndRefresh: jest.fn(),
-        resetUpdateSubscriptionPlan: jest.fn(),
-      };
-
-      function runTests(
-        plan: Plan,
-        expectedMsgId: string,
-        expectedMsg: string
-      ) {
-        const props = {
-          ...baseProps,
-          selectedPlan: plan,
-        };
-        const testRenderer = TestRenderer.create(
-          <SubscriptionUpgrade {...props} />
-        );
-        const testInstance = testRenderer.root;
-        const legalCheckbox = testInstance.findByProps({ id: expectedMsgId });
-        const expectedAmount = getLocalizedCurrency(plan.amount, plan.currency);
-
-        expect(legalCheckbox.props.vars.amount).toStrictEqual(expectedAmount);
-        expect(legalCheckbox.props.vars.intervalCount).toBe(
-          plan.interval_count
-        );
-        expect(legalCheckbox.props.children.props.children).toBe(expectedMsg);
-      }
-
-      it('renders Localized for daily plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_daily';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-day';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 daily</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for 6 days plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_6days';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-day';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 days</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for weekly plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_weekly';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-week';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 weekly</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for 6 weeks plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_6weeks';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-week';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 weeks</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for monthly plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_monthly';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-month';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 monthly</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for 6 months plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_6months';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-month';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 months</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for yearly plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_yearly';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-year';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 yearly</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-
-      it('renders Localized for years plan with correct props and displays correct default string', async () => {
-        const plan_id = 'plan_6years';
-        const plan = findMockPlan(plan_id);
-        const expectedMsgId = 'sub-update-confirm-with-legal-links-year';
-        const expectedMsg =
-          'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 years</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-        runTests(plan, expectedMsgId, expectedMsg);
-      });
-    });
-
-    describe('Fluent Localized Text', () => {
-      const bundle = setupFluentLocalizationTest('en-US');
-      const amount = getLocalizedCurrency(500, 'USD');
-      const args = {
-        amount,
-      };
-
-      describe('when the localized id is sub-update-confirm-with-legal-links-day', () => {
-        const msgId = 'sub-update-confirm-with-legal-links-day';
-
-        it('returns the correct string for an interval count of 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 daily</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 1,
-          });
-          expect(actual).toEqual(expected);
-        });
-
-        it('returns the correct string for an interval count greater than 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 days</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 6,
-          });
-          expect(actual).toEqual(expected);
-        });
-      });
-
-      describe('when the localized id is sub-update-confirm-with-legal-links-week', () => {
-        const msgId = 'sub-update-confirm-with-legal-links-week';
-
-        it('returns the correct string for an interval count of 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 weekly</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 1,
-          });
-          expect(actual).toEqual(expected);
-        });
-
-        it('returns the correct string for an interval count greater than 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 weeks</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 6,
-          });
-          expect(actual).toEqual(expected);
-        });
-      });
-
-      describe('when the localized id is sub-update-confirm-with-legal-links-month', () => {
-        const msgId = 'sub-update-confirm-with-legal-links-month';
-
-        it('returns the correct string for an interval count of 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 monthly</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 1,
-          });
-          expect(actual).toEqual(expected);
-        });
-
-        it('returns the correct string for an interval count greater than 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 months</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 6,
-          });
-          expect(actual).toEqual(expected);
-        });
-      });
-
-      describe('when the localized id is sub-update-confirm-with-legal-links-year', () => {
-        const msgId = 'sub-update-confirm-with-legal-links-year';
-
-        it('returns the correct string for an interval count of 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 yearly</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 1,
-          });
-          expect(actual).toEqual(expected);
-        });
-
-        it('returns the correct string for an interval count greater than 1', () => {
-          const expected =
-            'I authorize Mozilla, maker of Firefox products, to charge my payment method <strong>$5.00 every 6 years</strong>, according to <termsOfServiceLink>Terms of Service</termsOfServiceLink> and <privacyNoticeLink>Privacy Notice</privacyNoticeLink>, until I cancel my subscription.';
-
-          const actual = getLocalizedMessage(bundle, msgId, {
-            ...args,
-            intervalCount: 6,
-          });
-          expect(actual).toEqual(expected);
-        });
-      });
-    });
   });
 });
